#feat 회원 정보수정 유효성 검사 추가
## 🏷 작업 분류

- [x]  기능 추가
- [ ]  리팩토링
- [ ]  버그 수정
- [ ]  환경 수정 및 기타

## 💡 작업 개요
회원 정보수정에서 유효성검사 추가하고 닉네임은 변경은 삭제하였습니다.
## ***📜** 리뷰 요청사항*
